work on relation, relation list
